+++ src/main/scala/fr/janalyse/ssh/Expect.scala
@@ -1,0 +1,7 @@
+package fr.janalyse.ssh
+
+case class Expect(
+  when: (String) => Boolean,
+  send: String
+  )
+